daemon/requirements.txt
psutil>=5.9.0
httpx>=0.27.0
pyyaml>=6.0
